+++ src/main/java/org/opentripplanner/model/Station.java
@@ -127,7 +127,7 @@ public class Station extends TransitEntity implements StopCollection {
    * adding adjusting the cost for all board-/alight- events in the routing search.
    * <p/>
    * To not interfere with request parameters this must be implemented in a neutral way. This mean
-   * that the {@link StopTransferPriority#ALLOWED} (witch is default) should a nett-effect of
+   * that the {@link StopTransferPriority#ALLOWED} (which is default) should a nett-effect of
    * adding 0 - zero cost.
    */
   public StopTransferPriority getPriority() {