+++ src/main/java/org/opentripplanner/routing/algorithm/raptor/router/TransitRouter.java
@@ -1,12 +1,12 @@
 package org.opentripplanner.routing.algorithm.raptor.router;
 
+import java.time.Duration;
 import java.time.ZonedDateTime;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
-import org.opentripplanner.ext.dataoverlay.routing.DataOverlayContext;
 import org.opentripplanner.model.plan.Itinerary;
 import org.opentripplanner.routing.algorithm.mapping.RaptorPathToItineraryMapper;
 import org.opentripplanner.routing.algorithm.raptor.router.street.AccessEgressRouter;
@@ -27,10 +27,10 @@ import org.opentripplanner.routing.api.response.InputField;
 import org.opentripplanner.routing.api.response.RoutingError;
 import org.opentripplanner.routing.api.response.RoutingErrorCode;
 import org.opentripplanner.routing.error.RoutingValidationException;
+import org.opentripplanner.routing.fares.FareService;
 import org.opentripplanner.routing.framework.DebugTimingAggregator;
 import org.opentripplanner.routing.graph.GraphIndex;
 import org.opentripplanner.routing.graph.Vertex;
-import org.opentripplanner.routing.fares.FareService;
 import org.opentripplanner.standalone.server.Router;
 import org.opentripplanner.transit.raptor.RaptorService;
 import org.opentripplanner.transit.raptor.api.path.Path;
@@ -47,18 +47,22 @@ public class TransitRouter {
     private final RoutingRequest request;
     private final Router router;
     private final DebugTimingAggregator debugTimingAggregator;
-    private final ZonedDateTime searchStartTime;
+    private final ZonedDateTime transitSearchTimeZero;
+    private final AdditionalSearchDays additionalSearchDays;
+
 
     private TransitRouter(
             RoutingRequest request,
             Router router,
-            ZonedDateTime searchStartTime,
+            ZonedDateTime transitSearchTimeZero,
+            AdditionalSearchDays additionalSearchDays,
             DebugTimingAggregator debugTimingAggregator
     ) {
         this.request = request;
         this.router = router;
+        this.transitSearchTimeZero = transitSearchTimeZero;
+        this.additionalSearchDays = additionalSearchDays;
         this.debugTimingAggregator = debugTimingAggregator;
-        this.searchStartTime = searchStartTime;
     }
 
     private TransitRouterResult route() {
@@ -66,7 +70,7 @@ public class TransitRouter {
             return new TransitRouterResult(List.of(), null);
         }
 
-        if (!router.graph.transitFeedCovers(request.getDateTimeOriginalSearch())) {
+        if (!router.graph.transitFeedCovers(request.getDateTime())) {
             throw new RoutingValidationException(List.of(
                     new RoutingError(RoutingErrorCode.OUTSIDE_SERVICE_PERIOD, InputField.DATE_TIME)
             ));
@@ -94,7 +98,7 @@ public class TransitRouter {
         // Prepare transit search
         var raptorRequest = RaptorRequestMapper.mapRequest(
                 request,
-                searchStartTime,
+                transitSearchTimeZero,
                 accessEgresses.getAccesses(),
                 accessEgresses.getEgresses()
         );
@@ -126,7 +130,7 @@ public class TransitRouter {
         RaptorPathToItineraryMapper itineraryMapper = new RaptorPathToItineraryMapper(
                 router.graph,
                 transitLayer,
-                searchStartTime,
+                transitSearchTimeZero,
                 request
         );
         FareService fareService = router.graph.getService(FareService.class);
@@ -219,6 +223,7 @@ public class TransitRouter {
             if (OTPFeature.FlexRouting.isOn() && mode == StreetMode.FLEXIBLE) {
                 var flexAccessList = FlexAccessEgressRouter.routeAccessEgress(
                         accessRequest,
+                        additionalSearchDays,
                         router.routerConfig.flexParameters(request),
                         isEgress
                 );
@@ -235,15 +240,16 @@ public class TransitRouter {
     ) {
         var graph = router.graph;
 
+
         try (RoutingRequest transferRoutingRequest = Transfer.prepareTransferRoutingRequest(request)) {
             transferRoutingRequest.setRoutingContext(graph, (Vertex) null, null);
 
             return new RaptorRoutingRequestTransitData(
                     graph.getTransferService(),
                     transitLayer,
-                    searchStartTime,
-                    request.arriveBy ? request.additionalSearchDaysBeforeToday : 0,
-                    request.arriveBy ? 0 : request.additionalSearchDaysAfterToday,
+                    transitSearchTimeZero,
+                    additionalSearchDays.additionalSearchDaysInPast(),
+                    additionalSearchDays.additionalSearchDaysInFuture(),
                     createRequestTransitDataProviderFilter(graph.index),
                     transferRoutingRequest
             );
@@ -291,9 +297,10 @@ public class TransitRouter {
     public static TransitRouterResult route(
             RoutingRequest request,
             Router router,
-            ZonedDateTime searchStartTime,
+            ZonedDateTime transitSearchTimeZero,
+            AdditionalSearchDays additionalSearchDays,
             DebugTimingAggregator debugTimingAggregator
     ) {
-        return new TransitRouter(request, router, searchStartTime, debugTimingAggregator).route();
+        return new TransitRouter(request, router, transitSearchTimeZero, additionalSearchDays, debugTimingAggregator).route();
     }
 }