+++ src/main/java/org/opentripplanner/util/ProgressTracker.java
@@ -19,7 +19,7 @@ import java.util.function.Consumer;
  * 'timer' prevent notification unless a minimum amount of time is passed since last time the
  * caller was notified. The quiet period is set to 5 seconds.
  * <p>
- * There is also a 'minBlockSize' witch prevent the tracker of calling the
+ * There is also a 'minBlockSize' which prevent the tracker of calling the
  * {@link System#currentTimeMillis()} for each step, instead the timer is checked once for each
  * block of steps. This make the progress step up in regular, nice to read, chunks too.
  * <p>
@@ -78,7 +78,7 @@ public class ProgressTracker {
      * Track progress for the given action.
      * @param actionName the action name to include in the notification strings.
      * @param minBlockSize the minimum number of steps between each time check. A reasonably value
-     *                     is to use the number of steps witch would take approximately 1 second.
+     *                     is to use the number of steps which would take approximately 1 second.
      *                     Set this lower if the time variation for each step is big.
      * @param size The expected number the step method is called. If negative
      *             the size is considered unknown.
@@ -92,7 +92,7 @@ public class ProgressTracker {
      *
      * @param actionName the action name to include in the notification strings.
      * @param minBlockSize the minimum number of steps between each time check. A reasonably value
-     *                     is to use the number of steps witch would take approximately 1 second.
+     *                     is to use the number of steps which would take approximately 1 second.
      *                     Set this lower if the time variation for each step is big.
      * @param size The expected number the step method is called. If negative
      *             the size is considered unknown.
@@ -118,7 +118,7 @@ public class ProgressTracker {
      *
      * @param actionName the action name to include in the notification strings.
      * @param minBlockSize the minimum number of steps between each time check. A reasonably value
-     *                     is to use the number of steps witch would take approximately 1 second.
+     *                     is to use the number of steps which would take approximately 1 second.
      *                     Set this lower if the time variation for each step is big.
      * @param size The expected number the step method is called. If negative
      *             the size is considered unknown.