+++ src/test/java/org/opentripplanner/graph_builder/module/osm/TestIntermediatePlaces.java
@@ -190,9 +190,9 @@ public class TestIntermediatePlaces {
         Calendar arriveTime = Calendar.getInstance(timeZone);
         if (request.arriveBy) {
             departTime = itinerary.legs.get(0).getStartTime();
-            arriveTime.setTimeInMillis(request.getDateTimeOriginalSearch().toEpochMilli());
+            arriveTime.setTimeInMillis(request.getDateTime().toEpochMilli());
         } else {
-            departTime.setTimeInMillis(request.getDateTimeCurrentPage().toEpochMilli());
+            departTime.setTimeInMillis(request.getDateTime().toEpochMilli());
             arriveTime = itinerary.legs.get(itinerary.legs.size() - 1).getEndTime();
         }
         long sumOfDuration = 0;